+++ src/test/java/de/alpharogroup/enums/CustomizableEnumTest.java
@@ -35,9 +35,9 @@ import org.testng.annotations.Test;
 
 import de.alpharogroup.test.objects.enums.Brands;
 import de.alpharogroup.test.objects.enums.Gender;
-import de.alpharogroup.test.objects.evaluations.EqualsEvaluator;
-import de.alpharogroup.test.objects.evaluations.HashcodeEvaluator;
-import de.alpharogroup.test.objects.evaluations.ToStringEvaluator;
+import de.alpharogroup.evaluate.object.EqualsEvaluator;
+import de.alpharogroup.evaluate.object.HashcodeEvaluator;
+import de.alpharogroup.evaluate.object.ToStringEvaluator;
 
 /**
  * The unit test class for the class {@link CustomizableEnum}.