+++ src/test/java/de/alpharogroup/file/namefilter/SimpleFilenameFilterTest.java
@@ -34,7 +34,7 @@ import java.net.URISyntaxException;
 import org.testng.annotations.Test;
 
 import de.alpharogroup.lang.ClassExtensions;
-import de.alpharogroup.test.objects.evaluations.ToStringEvaluator;
+import de.alpharogroup.evaluate.object.ToStringEvaluator;
 
 /**
  * The unit test class for the class {@link SimpleFilenameFilter}.