+++ src/test/java/de/alpharogroup/lang/model/PropertiesKeyAndParametersTest.java
@@ -35,8 +35,8 @@ import org.meanbean.test.ConfigurationBuilder;
 import org.testng.annotations.Test;
 
 import de.alpharogroup.meanbean.factories.ObjectArrayFactory;
-import de.alpharogroup.test.objects.evaluations.EqualsHashCodeAndToStringEvaluator;
-import de.alpharogroup.test.objects.evaluations.SilentEqualsHashCodeAndToStringEvaluator;
+import de.alpharogroup.evaluate.object.EqualsHashCodeAndToStringEvaluator;
+import de.alpharogroup.evaluate.object.SilentEqualsHashCodeAndToStringEvaluator;
 
 /**
  * The unit test class for the class {@link PropertiesKeyAndParameters}.